Rename directory to octrace.
